requirements.txt
requests==2.22.0
selenium==3.141.0
pandas==1.1.3
numpy==1.19.2
requests_cache==0.5.2
scikit_learn==0.24.2
python-dotenv==0.18.0
